refactor: 쿼리 스트링 타입 변경

+++ src/main/java/smw/capstone/controller/BoardController.java
@@ -38,7 +38,7 @@ public class BoardController {
      * 게시물 삭제
      */
     @PostMapping("/delete")
-    public ResponseEntity<String> delBoard(int boardId /*사용자 정보*/) {
+    public ResponseEntity<String> delBoard(Long boardId /*사용자 정보*/) {
         boardService.deleteBoard(boardId);
         return ResponseEntity.ok().body("게시물이 삭제되었습니다.");
     }
